- get student works

+++ src/main/java/com/example/meditatii/services/StudentService.java
@@ -1,7 +1,6 @@
 package com.example.meditatii.services;
 
 import com.example.meditatii.DTO.StudentDTO;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
@@ -9,6 +8,6 @@ import java.util.Optional;
 public interface StudentService {
     List<StudentDTO> listAllStudents();
     StudentDTO saveStudent(StudentDTO studentDTO);
-    Optional<StudentDTO> getStudentById(final Long id);
+    Optional<StudentDTO> getStudentById(final Integer id);
     void updateStudent(final StudentDTO studentDTO);
 }
