+++ src/main/java/com/example/meditatii/services/StudentServiceJPA.java
@@ -24,9 +24,9 @@ public class StudentServiceJPA implements StudentService{
                 .collect(Collectors.toList());
     }
 
-    public Optional<StudentDTO> getStudentById(final Long id){
+    public Optional<StudentDTO> getStudentById(final Integer id){
         return Optional.ofNullable
-                (mapper.studentToStudentDto(repository.findById(id).orElse(null)));
+                (mapper.studentToStudentDto(repository.findById(Long.valueOf(id)).orElse(null)));
     }
